+++ app/src/main/java/com/example/banchan/presentation/adapter/home/CommonAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.banchan.domain.model.ItemModel
 import com.example.banchan.presentation.home.maincook.Filter
-import com.example.banchan.presentation.home.maincook.adapter.MediumMenuViewHolder
+import com.example.banchan.presentation.adapter.main.MediumMenuViewHolder
 
 class CommonAdapter(private val onFilterChanged: (Filter) -> Unit) :
     ListAdapter<CommonItemListModel, RecyclerView.ViewHolder>(DiffCallback()) {
