+++ app/src/main/java/com/example/banchan/presentation/adapter/main/MainAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.banchan.presentation.home.maincook.adapter
+package com.example.banchan.presentation.adapter.main
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
