+++ app/src/main/java/com/example/banchan/presentation/home/maincook/MainCookFragment.kt
@@ -10,9 +10,9 @@ import com.example.banchan.R
 import com.example.banchan.databinding.FragmentMainCookBinding
 import com.example.banchan.domain.model.ItemListModel
 import com.example.banchan.presentation.base.BaseFragment
-import com.example.banchan.presentation.home.maincook.adapter.MainAdapter
-import com.example.banchan.presentation.home.maincook.adapter.SpacingItemDecorator
-import com.example.banchan.presentation.home.maincook.adapter.Type
+import com.example.banchan.presentation.adapter.main.MainAdapter
+import com.example.banchan.presentation.adapter.main.SpacingItemDecorator
+import com.example.banchan.presentation.adapter.main.Type
 import com.example.banchan.util.dimen.dpToPx
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
