+++ app/src/main/java/com/example/banchan/presentation/home/side/SideFragment.kt
@@ -8,8 +8,8 @@ import androidx.recyclerview.widget.GridLayoutManager
 import com.example.banchan.R
 import com.example.banchan.databinding.FragmentSoupBinding
 import com.example.banchan.presentation.base.BaseFragment
-import com.example.banchan.presentation.home.maincook.adapter.MainAdapter
-import com.example.banchan.presentation.home.maincook.adapter.SpacingItemDecorator
+import com.example.banchan.presentation.adapter.main.MainAdapter
+import com.example.banchan.presentation.adapter.main.SpacingItemDecorator
 import com.example.banchan.presentation.adapter.home.CommonAdapter
 import com.example.banchan.util.dimen.dpToPx
 import dagger.hilt.android.AndroidEntryPoint
